Move wordpress components to src/components
