+++ website/src/components/index.ts
@@ -4,3 +4,4 @@ export { Select, Label } from "./select"
 export { AudioPlayer } from "./audio-player"
 export { Breadcrumbs } from "./breadcrumbs"
 export { Carousel } from "./carousel"
+export { default as WordpressPage, WordpressPageContents } from "./wordpress"
