+++ website/src/pages/index.page.tsx
@@ -1,8 +1,7 @@
 import React from "react"
 import { Helmet } from "react-helmet"
-import { Carousel, Link } from "src/components"
+import { Carousel, Link, WordpressPage } from "src/components"
 import { fullWidth, paddedCenterColumn } from "src/style/utils.css"
-import WordpressPage from "src/wordpress-page"
 import * as Dailp from "../graphql/dailp"
 import Layout from "../layout"
 import { collectionRoute } from "../routes"
