+++ website/src/pages/wordpress.page.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import { Helmet } from "react-helmet"
+import { WordpressPageContents } from "src/components"
 import * as Wordpress from "src/graphql/wordpress"
 import { edgePadded, fullWidth } from "src/style/utils.css"
-import { WordpressContents } from "src/wordpress-page"
 import Layout from "../layout"
 import { wordpressUrl } from "../theme.css"
 
@@ -38,7 +38,7 @@ const Contents = (props: { slug: string }) => {
         <header>
           <h1>{page.title}</h1>
         </header>
-        <WordpressContents content={page.content ?? ""} />
+        <WordpressPageContents content={page.content} />
       </>
     )
   } else if (fetching) {
